requirements.txt
simpy>=4.1
networkx>=3.2
numpy>=1.26
pandas>=2.1
PyYAML>=6.0
matplotlib>=3.8
seaborn>=0.13
python-docx>=1.1
